Remarks from review Implemented
