+++ src/keystore/error.rs
@@ -1,7 +1,6 @@
 //! # Keystore files (UTC / JSON) module errors
 
-use rustc_serialize::json;
-use std::{error, fmt, io};
+use std::{error, fmt};
 
 /// Keystore file errors
 #[derive(Debug)]
@@ -14,10 +13,6 @@ pub enum Error {
     UnsupportedPrf(String),
     /// `keccak256_mac` field validation failed
     FailedMacValidation,
-    ///
-    FileCreation,
-    ///
-    InvalidEncoding,
 }
 
 impl fmt::Display for Error {
@@ -31,8 +26,6 @@ impl fmt::Display for Error {
                 write!(f, "Unsupported pseudo-random function: {}", str)
             }
             Error::FailedMacValidation => write!(f, "Message authentication code failed"),
-            Error::FileCreation => write!(f, "Can't create file for KeyFile"),
-            Error::InvalidEncoding => write!(f, "Can't encode KeyFile"),
         }
     }
 }
@@ -46,17 +39,5 @@ impl error::Error for Error {
         match *self {
             _ => None,
         }
-    }
-}
-
-impl From<json::EncoderError> for Error {
-    fn from(_: json::EncoderError) -> Self {
-        Error::InvalidEncoding
-    }
-}
-
-impl From<io::Error> for Error {
-    fn from(_: io::Error) -> Self {
-        Error::FileCreation
     }
 }
